refactor: 상수 관련 항목 디렉토리 분리
