+++ src/main/java/lotto/constant/LottoConstant.java
@@ -1,0 +1,7 @@
+package lotto.constant;
+
+public class LottoConstant {
+    public static final Integer START_INDEX = 1;
+    public static final Integer END_INDEX = 45;
+    public static final Integer PICK_COUNT = 6;
+}
