+++ src/main/java/lotto/constant/ViewConstant.java
@@ -1,4 +1,4 @@
-package lotto.ui;
+package lotto.constant;
 
 public class ViewConstant {
     public static final String INPUT_MONEY_MESSAGE = "구입금액을 입력해 주세요.";
